+++ askbot/models/post.py
@@ -3,7 +3,6 @@ import datetime
 import operator
 import logging
 
-from django.utils.html import strip_tags
 from django.contrib.sitemaps import ping_google
 from django.utils import html
 from django.conf import settings as django_settings
@@ -421,26 +420,8 @@ class Post(models.Model):
         removed_mentions - list of mention <Activity> objects - for removed ones
         """
 
-        if self.post_type in ('question', 'answer', 'tag_wiki', 'reject_reason'):
-            _urlize = False
-            _use_markdown = (askbot_settings.EDITOR_TYPE == 'markdown')
-        elif self.is_comment():
-            _urlize = True
-            _use_markdown = (askbot_settings.EDITOR_TYPE == 'markdown')
-        else:
-            raise NotImplementedError
-
-        text = self.text
-
-        if _urlize:
-            text = html.urlize(text)
-
-        if _use_markdown:
-            text = sanitize_html(markup.get_parser().convert(text))
-
-        if askbot_settings.EDITOR_TYPE == 'tinymce':
-            #todo: see what can be done with the "object" tag
-            text = strip_tags(text, ['script', 'style', 'link'])
+        text_converter = self.get_text_converter()
+        text = text_converter(self.text)
 
         #todo, add markdown parser call conditional on
         #self.use_markdown flag
@@ -613,6 +594,23 @@ class Post(models.Model):
             return None
 
         return answer
+
+    def get_text_converter(self):
+        have_simple_comment = (
+            self.is_comment() and 
+            askbot_settings.COMMENTS_EDITOR_TYPE == 'plain-text'
+        )
+        if have_simple_comment:
+            parser_type = 'plain-text'
+        else:
+            parser_type = askbot_settings.EDITOR_TYPE
+
+        if parser_type == 'plain-text':
+            return markup.plain_text_input_converter
+        elif parser_type == 'markdown':
+            return markup.markdown_input_converter
+        elif parser_type == 'tynymce':
+            return markup.tinymce_input_converter
 
     def has_group(self, group):
         """true if post belongs to the group"""