+++ askbot/doc/source/changelog.rst
@@ -3,6 +3,8 @@ Changes in Askbot
 
 Development version
 -------------------
+* Added possibility to choose editor for comments: plain text, or same as
+  editor used for the questions or answers: WMD or TinyMCE.
 * Added a placeholder template for the custom javascript on the question page
 * Allowed to disable the big "ask" button.
 * Some support for the media compression (Tyler Mandry)